# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Saúde UNASP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/imc/calcular": {
            "post": {
                "tags": [
                    "IMC"
                ],
                "summary": "Faz o cálculo do IMC",
                "description": "Realiza o cálculo do imc",
                "operationId": "7b1f15047a232839d9b2495ef0a49aa2",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "applicationJson": {
                                "summary": "application/json",
                                "value": "application/json"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "peso": "70.2",
                                    "altura": "1.71"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Ok",
                                        "value": {
                                            "message": "Seu IMC é 20.2 Está classificado como Normal"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Parâmetros enviados são inválidos",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Entidade Não processável",
                                        "value": [
                                            {
                                                "message": "O campo \"peso\" é requerido. (and 1 more error)",
                                                "errors": {
                                                    "peso": [
                                                        "O campo \"peso\" é requerido."
                                                    ],
                                                    "altura": [
                                                        "O campo \"altura\" é requerido."
                                                    ]
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/endereco/{cep}": {
            "get": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Constulta endereço pelo CEP",
                "description": "Constulta endereço pelo CEP",
                "operationId": "b0ec92fde869651b6887a3936921829f",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "applicationJson": {
                                "summary": "application/json",
                                "value": "application/json"
                            }
                        }
                    },
                    {
                        "name": "cep",
                        "in": "path",
                        "description": "CEP",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "cepValido": {
                                "summary": "CEP Válido",
                                "value": "01001000"
                            },
                            "cepInvalido": {
                                "summary": "CEP Inválido",
                                "value": "0100100"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Endereco Encontrado",
                                        "value": {
                                            "cep": "01001-000",
                                            "logradouro": "Praça da Sé",
                                            "complemento": "lado ímpar",
                                            "bairro": "Sé",
                                            "localidade": "São Paulo",
                                            "uf": "SP",
                                            "ibge": "3550308",
                                            "gia": "1004",
                                            "ddd": "11",
                                            "siafi": "7107"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "CEP Inválido",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "CEP Inválido",
                                        "value": {
                                            "message": "CEP deve ter 8 dígitos"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CEP não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "CEP não encontrado",
                                        "value": {
                                            "message": "Deu algo errado na busca deste endereço. Verifique se digitou o CEP corretamente."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Erro Interno",
                                        "value": {
                                            "message": "Ocorreu um erro inesperado, por favor, tente mais tarde.",
                                            "error": ""
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/endereco/estados": {
            "get": {
                "tags": [
                    "Endereco"
                ],
                "summary": "Lista de estados brasileiros",
                "description": "Lista de estados brasileiros",
                "operationId": "1f4d9d1d3ff43648bfed27717297b794",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Accept",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "applicationJson": {
                                "summary": "application/json",
                                "value": "application/json"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Lista de Estados",
                                        "value": [
                                            {
                                                "uf": "AC",
                                                "nome": "Acre"
                                            },
                                            {
                                                "uf": "AL",
                                                "nome": "Alagoas"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Sem conteúdo"
                    }
                }
            }
        }
    },
...
}